requirements.txt
django==5.0.6
gunicorn==23.0.0
djangorestframework
